A Bin's ceiling comes from the Ruleset in force and its level is a long, and the ADR that asked for the first was wrong about the code
Hash-moving queue item 3 (plans/0003), and the queue's last. adr/0064 and
adr/0065 in one commit rather than two, which inverts this queue's usual rule
for the reason the rule exists: they change the same two columns — one the
declaration, one the width — so a baseline that moved for both is attributable
to neither if they ship apart.

adr/0064. Bins.Capacity is Rows.Derived<long>, not a saved column. World.CreateBin
no longer takes a ceiling; it derives one from (Building kind, Resource) against
the Ruleset in force. World.RebuildCapacities runs in RebuildDerived and inside
Adopt, so a retuned capacity reaches the Buildings already standing rather than
only the next one raised. A Bin left above a lowered ceiling drains rather than
clamps: headroom goes negative, IntegerMath.FloorDiv rounds toward negative
infinity so a producer's affordable count falls below its floor and it stops,
a consumer is untouched, and the stock comes back into range by being spent.
Clamping would destroy Goods on a keystroke.

adr/0065. A Bin's level and capacity are long, and so is the whole write path —
BinTable.Move, World.Deposit/Withdraw, LevelAt, HeadroomAt, RuleVerdict.Applications,
RuleEngine.Check/Walk/Band/Requirement/Touch, Readouts.Read, Violation.Other and
the crash artifact's parser. Widening the column alone would have moved the
narrowing to the argument. Money is a long and adr/0031 puts money in a Bin, so
every payment into one was narrowing 64 bits to 32.

The end-of-run check adr/0064 obliges is Invariant.BinCapacityMatchesItsDeclaration,
id 29, beside NoWaiterSleepsOnANonBlockingBin. A derived column nobody rebuilt is
stale, and stale is silent: every row stays self-consistent and the only wrong
thing is a ceiling that agrees with a Ruleset nobody is running.

Three findings.

An ADR was wrong about the code, and the cause is a missing test. adr/0064
recorded a live defect on the ground that RulesetLoader refused nothing for a
duplicate (kind, Resource) Bin declaration. It has refused it since slice 7
task 8 — and that refusal was the one guard in the loader with no test, so the
suite you read to find out what the loader refuses did not name it. plans/0012
Cause 1 on a different axis: not a second copy of a fact drifting from the
first, but a fact with no copy at all, re-derived wrongly from the shape of its
absence. The ADR is amended and the test ships here.

world-hash.txt did not move, again. This change deleted a column from the fold
and doubled another's width, and only session-trace.txt noticed, because
GoldenFixtures.Build() holds no Bins at all. Three of the four queue items moved
exactly one of the two committed baselines and the same one each time.

long reached a boundary neither ADR named. RuleEngine.Emit's amount x applications
crosses into a Map Layer cell that is an int. That is now a loud throw with the
number in the message rather than a silent cast: the whole write path includes
the paths that leave.

Records in plans/0018 (Tasks 3 and 4's implementation record), plans/0003 (the
queue is empty), plans/0000-board.md, CLAUDE.md, CONTEXT.md, docs/04 and a
forward note on adr/0031, whose "prefer an explicit unbounded to a large
sentinel" adr/0065 withdraws as unachievable under lint 7.

dotnet test green at 839. session-trace.txt re-recorded; world-hash.txt unchanged.

+++ src/Borough.Core/Invariants/InvariantRegistry.cs
@@ -157,7 +157,7 @@ public sealed class InvariantRegistry
     /// <param name="invariant">Which invariant, if it is not.</param>
     /// <param name="slot">The row being written.</param>
     /// <param name="other">A second row involved, where there is one.</param>
-    public void Require(bool held, Invariant invariant, int slot = -1, int other = -1)
+    public void Require(bool held, Invariant invariant, int slot = -1, long other = -1)
     {
         if (!held)
         {
@@ -167,7 +167,7 @@ public sealed class InvariantRegistry
 
     /// <summary>Records or throws, according to <see cref="Collect"/>.</summary>
     /// <exception cref="InvariantViolationException">A violation, and the run is not collecting.</exception>
-    public void Report(Invariant invariant, int slot = -1, int other = -1)
+    public void Report(Invariant invariant, int slot = -1, long other = -1)
     {
         var violation = new Violation(invariant, Tick, slot, other);
